refactor: move useHardwareBack into the framework module
