+++ src/features/notifications/screens/NotificationPermissionScreen.tsx
@@ -14,7 +14,7 @@ import mockNotificationsIOS from '@/assets/mockNotificationsIOS.png';
 import backgroundImage from '@/assets/notificationsPromoSheetBackground.png';
 import ButtonPressAnimation from '@/components/animations/ButtonPressAnimation';
 import { Bleed, Box, ColorModeProvider, Column, Columns, Stack, Text } from '@/design-system';
-import { useHardwareBackOnFocus } from '@/hooks/useHardwareBack';
+import { useHardwareBackOnFocus } from '@/framework/ui/hooks/useHardwareBack';
 import * as i18n from '@/languages';
 import Navigation from '@/navigation/Navigation';
 import Routes from '@/navigation/routesNames';
