+++ src/screens/QRScannerScreen.tsx
@@ -12,9 +12,9 @@ import { Centered } from '@/components/layout';
 import { Navbar } from '@/components/navbar/Navbar';
 import { CameraDimmer, QRCodeScanner } from '@/components/qrcode-scanner';
 import { AccentColorProvider, Box, ColorModeProvider, Text } from '@/design-system';
+import { useHardwareBack } from '@/framework/ui/hooks/useHardwareBack';
 import styled from '@/framework/ui/styled-thing';
 import useDimensions from '@/hooks/useDimensions';
-import useHardwareBack from '@/hooks/useHardwareBack';
 import { useIsForeground } from '@/hooks/useIsForeground';
 import useScanner from '@/hooks/useScanner';
 import * as i18n from '@/languages';
