+++ src/screens/WelcomeScreen/WelcomeScreen.tsx
@@ -21,9 +21,9 @@ import { useSafeAreaInsets } from 'react-native-safe-area-context';
 import { analytics } from '@/analytics';
 import { Box } from '@/design-system';
 import { IS_TEST } from '@/env';
+import { useHardwareBackOnFocus } from '@/framework/ui/hooks/useHardwareBack';
 import { opacity } from '@/framework/ui/utils/opacity';
 import { WalletLoadingStates } from '@/helpers/walletLoadingStates';
-import { useHardwareBackOnFocus } from '@/hooks/useHardwareBack';
 import { hideSplashScreen } from '@/hooks/useHideSplashScreen';
 import * as i18n from '@/languages';
 import { ensureError, logger, RainbowError } from '@/logger';
